// app/__generated__/IssueLabels_issue.graphql.ts
export type IssueLabels_issue$data = {
  readonly id: string;
  readonly issue_labelsCollection: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly labels: {
          readonly color: string;
          readonly id: string;
          readonly name: string;
          readonly nodeId: string;
        } | null | undefined;
        readonly nodeId: string;
      };
    }>;
  } | null | undefined;
  readonly nodeId: string;
  readonly " $fragmentType": "IssueLabels_issue";
};
export type IssueLabels_issue$key = {
  readonly " $data"?: IssueLabels_issue$data;
  readonly " $fragmentSpreads": FragmentRefs<"IssueLabels_issue">;
};